refactor: move find functions
